+++ docs/changelog.rst
@@ -7,6 +7,8 @@ Changelog
 24.10.3
 =======
 - Add :ref:`ASYNC124 <async124>` async-function-could-be-sync
+- :ref:`ASYNC91x <ASYNC910>` now correctly handles ``await()`` in parameter lists.
+- Fixed a bug with :ref:`ASYNC91x <ASYNC910>` and nested empty functions.
 
 24.10.2
 =======